# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: ContentFlow API
  description: AI SEO content generation platform.
  version: 0.1.0
servers:
  - url: /
    description: Same origin as the docs (works on any host/port)
  - url: https://multiagent-seo.8cell.tech
    description: Staging
...
paths:
...
  /users:
    get:
      operationId: listUsers
      security:
        - bearerAuth: [ admin ]
      tags: [ Auth ]
      summary: List users
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /api-tokens:
    post:
      operationId: createApiToken
      security:
        - bearerAuth: [ admin ]
      tags: [ API Tokens ]
      summary: Create an API token (the secret is returned only once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiTokenRequest"
      responses:
        "201":
          description: Created — save the token now, it can't be retrieved later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTokenCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listApiTokens
      security:
        - bearerAuth: [ admin ]
      tags: [ API Tokens ]
      summary: List your API tokens (metadata only)
      responses:
        "200":
          description: Your active tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiToken"
  /api-tokens/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: deleteApiToken
      security:
        - bearerAuth: [ admin ]
      tags: [ API Tokens ]
      summary: Revoke an API token
      responses:
        "204":
          description: Revoked
        "404":
          $ref: "#/components/responses/NotFound"
...
  /my/cases:
    get:
      operationId: listMyCases
      security:
        - bearerAuth: [ advocate, admin ]
      tags: [ My ]
      summary: The cases assigned to the advocate who is logged in — client contact, fee, what is paid, what is still owed, and the installments behind it
      responses:
        "200":
          description: Own cases, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyCaseList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /my/cases/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: setMyCaseStatus
      security:
        - bearerAuth: [ advocate, admin ]
      tags: [ My ]
      summary: Close an own case or put it back in progress; cancelling writes off what the client owes and stays an admin action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMyCaseStatusRequest"
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /my/clients:
    get:
      operationId: listMyClients
      security:
        - bearerAuth: [ advocate, admin ]
      tags: [ My ]
      summary: Clients the advocate has a case with — a client who only ever had a consultation is not here, consultations carry no advocate
      responses:
        "200":
          description: Own clients, most recent case first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyClientList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /my/clients/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getMyClient
      security:
        - bearerAuth: [ advocate, admin ]
      tags: [ My ]
      summary: One own client — name, phone, the advocate's own cases, the client's consultations and the note log; no address, birthdate or tax id, and no firm-wide revenue
      responses:
        "200":
          description: Client card, narrowed to what an advocate may see
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyClientCard"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such client, or none the advocate has a case with — the same answer either way, so an id cannot be probed for existence
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /my/clients/{id}/notes:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: addMyClientNote
      security:
        - bearerAuth: [ advocate, admin ]
      tags: [ My ]
      summary: Log a call or contact on an own client's card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddClientNoteRequest"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientNote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /my/settlement:
    get:
      operationId: getMySettlement
      security:
        - bearerAuth: [ advocate, admin ]
      tags: [ My ]
      summary: The advocate's own money — collected, their percentage, accrued, paid out, still owed, month by month; no colleague's row and no firm total
      responses:
        "200":
          description: Own settlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MySettlement"
        "403":
          $ref: "#/components/responses/Forbidden"
  /my/stats:
    get:
      operationId: getMyStats
      security:
        - bearerAuth: [ advocate, admin ]
      tags: [ My ]
      summary: Own scoreboard — cases by status, average fee, what clients still owe on own cases, collections by month
      responses:
        "200":
          description: Own statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyStats"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: The logged-in role may not use this operation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    User:
      type: object
      required: [ id, email, role, createdAt ]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateApiTokenRequest:
      type: object
      required: [ name ]
      properties:
        name:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "required,min=1,max=100"
    ApiToken:
      type: object
      required: [ id, name, prefix, createdAt ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        createdAt:
          type: string
          format: date-time
    ApiTokenCreated:
      type: object
      required: [ id, name, prefix, token, createdAt ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        token:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      description: RFC 7807 problem details
      required: [ title, status ]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    ClientNote:
      type: object
      required: [ id, text, created_by, created_at ]
      properties:
        id:
          type: string
        text:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
...
    AddClientNoteRequest:
      type: object
      required: [ text ]
      properties:
        text:
          type: string
...
    MyCase:
      type: object
      required: [ id, client_id, client_name, client_phone, category, status, description, fee, paid, owed, created_at, payments ]
      properties:
        id:
          type: string
        client_id:
          type: string
        client_name:
          type: string
        client_phone:
          type: string
        category:
          type: string
        status:
          type: string
        description:
          type: string
        fee:
          type: number
          format: double
        paid:
          type: number
          format: double
        owed:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        payments:
          type: array
          items:
            $ref: "#/components/schemas/MyCasePayment"
    MyCasePayment:
      type: object
      required: [ id, amount, paid_at ]
      properties:
        id:
          type: string
        amount:
          type: number
          format: double
        paid_at:
          type: string
          format: date
    MyCaseList:
      type: object
      required: [ items, total_fee, total_paid, total_owed ]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MyCase"
        total_fee:
          type: number
          format: double
        total_paid:
          type: number
          format: double
        total_owed:
          type: number
          format: double
    MyClient:
      type: object
      required: [ id, name, phone, cases, fee, paid, owed, last_case_at ]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        cases:
          type: integer
        fee:
          type: number
          format: double
        paid:
          type: number
          format: double
        owed:
          type: number
          format: double
        last_case_at:
          type: string
          format: date-time
    MyClientList:
      type: object
      required: [ items ]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MyClient"
    MyClientCard:
      type: object
      required: [ client, cases, consultations, notes ]
      properties:
        client:
          $ref: "#/components/schemas/MyClient"
        cases:
          type: array
          items:
            $ref: "#/components/schemas/MyCase"
        consultations:
          type: array
          items:
            $ref: "#/components/schemas/MyConsultation"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/ClientNote"
    MyConsultation:
      type: object
      required: [ id, scheduled_at, price, status, case_note ]
      properties:
        id:
          type: string
        scheduled_at:
          type: string
          format: date-time
        price:
          type: number
          format: double
        status:
          type: string
        case_note:
          type: string
    MyMonthMoney:
      type: object
      required: [ month, collected, accrued ]
      properties:
        month:
          type: string
          description: YYYY-MM
        collected:
          type: number
          format: double
        accrued:
          type: number
          format: double
    MySettlement:
      type: object
      required: [ advocate_id, full_name, commission_percent, collected, accrued, paid, outstanding, months, paid_is_partial ]
      properties:
        advocate_id:
          type: string
        full_name:
          type: string
        commission_percent:
          type: number
          format: double
        collected:
          type: number
          format: double
        accrued:
          type: number
          format: double
        paid:
          type: number
          format: double
        outstanding:
          type: number
          format: double
        months:
          type: array
          items:
            $ref: "#/components/schemas/MyMonthMoney"
        paid_is_partial:
          type: boolean
          description: Paid counts only payouts booked to this advocate by the generator; lump sums entered by hand name nobody, so outstanding can be overstated
    MyStats:
      type: object
      required: [ cases, clients, by_status, fee_total, paid_total, client_debt, avg_fee, months ]
      properties:
        cases:
          type: integer
        clients:
          type: integer
        by_status:
          type: array
          items:
            $ref: "#/components/schemas/MyStatusCount"
        fee_total:
          type: number
          format: double
        paid_total:
          type: number
          format: double
        client_debt:
          type: number
          format: double
        avg_fee:
          type: number
          format: double
        months:
          type: array
          items:
            $ref: "#/components/schemas/MyMonthMoney"
        first_case_at:
          type: string
          format: date-time
        last_payment_at:
          type: string
          format: date-time
    MyStatusCount:
      type: object
      required: [ status, count ]
      properties:
        status:
          type: string
        count:
          type: integer
    SetMyCaseStatusRequest:
      type: object
      required: [ status ]
      properties:
        status:
          type: string
          enum: [ in_progress, completed ]